+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/settings/MapboxCarStorage.kt
@@ -4,15 +4,15 @@ import android.content.Context
 import android.content.SharedPreferences
 import androidx.car.app.CarContext
 
-private const val CAR_SETTINGS_PREFERENCES_NAME = "car_settings"
+private const val CAR_SETTINGS_PREFERENCES_NAME = "mapbox_car_settings"
 
-class CarSettingsStorage(
+class MapboxCarStorage internal constructor(
     val carContext: CarContext
 ) {
     private val sharedPreferences: SharedPreferences =
         carContext.getSharedPreferences(CAR_SETTINGS_PREFERENCES_NAME, Context.MODE_PRIVATE)
 
-    fun readSharedPref(key: String, defaultValue: Boolean) =
+    fun readSharedPref(key: String, defaultValue: Boolean): Boolean =
         sharedPreferences.getBoolean(key, defaultValue)
 
     fun writeSharedPref(key: String, value: Boolean) {